// backend/services/matching-service/src/types.ts
export interface LocationCoordinates {
    latitude: number;
    longitude: number;
}